Rename Sensor into BasicSensor.
