+++ core/CMakeLists.txt
@@ -30,8 +30,10 @@ set(SRC
     "${CMAKE_CURRENT_SOURCE_DIR}/src/TelemetryRecorder.cc"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Utilities.cc"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/AbstractSensor.cc"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/AbstractMotor.cc"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/AbstractController.cc"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/Sensor.cc"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/BasicSensors.cc"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/BasicMotors.cc"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Model.cc"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Engine.cc"
 )
