+++ core/include/jiminy/core/AbstractSensor.tpp
@@ -1,5 +1,5 @@
 #include "jiminy/core/Model.h"
-#include "jiminy/core/Engine.h" // Required to get access to MIN_TIME_STEP and MAX_TIME_STEP
+#include "jiminy/core/Engine.h" // MIN_TIME_STEP and MAX_TIME_STEP
 
 namespace jiminy
 {
